+++ src/demo/data-fetching/pages/DataFetching/index.tsx
@@ -2,14 +2,14 @@ import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 
 import { useService } from '@/di';
-import asyncData from '@/common/utils/asyncData';
+import getResolveValue from '@/common/utils/getResolveValue';
 import useRerender from '@/common/hooks/useRerender';
 import withSuspense from '@/common/hocs/withSuspense';
 import DataService from '../../services/DataService';
 
 function DataFetchingPage() {
   const dataService = useService(DataService);
-  const data = asyncData(dataService.getData());
+  const data = getResolveValue(dataService.getData());
   const reload = useRerender();
   return (
     <>
